+++ app/src/main/java/com/example/p2project/IdleGame/EarnThread.java
@@ -20,8 +20,8 @@ public class EarnThread extends Thread{
                 @Override
                 public void run()
                 {
-                    main.treatView.setText(String.valueOf(CurrencyTracker.treats));
-                    main.earnView.setText(String.valueOf(CurrencyTracker.curEarn()));
+                    main.treatView.setText(String.valueOf(CurrencyTracker.treats.intValue()));
+                    main.earnView.setText(String.valueOf(CurrencyTracker.curEarn().intValue()));
                 }
             });
         }
